+++ packages/cli/src/commands/gdx/index.ts
@@ -1,6 +1,7 @@
 import { Command } from "commander";
 import { commandGdxPull } from "./pull.js";
 import { commandGdxPush } from "./push.js";
+import { commandGdxDiff } from "./diff.js";
 
 export const commandGdx = new Command("gdx").description("gdx management").action(() => {
   console.log("gdx");
@@ -8,3 +9,4 @@ export const commandGdx = new Command("gdx").description("gdx management").actio
 
 commandGdx.addCommand(commandGdxPull);
 commandGdx.addCommand(commandGdxPush);
+commandGdx.addCommand(commandGdxDiff);