+++ packages/cli/src/types.ts
@@ -1,7 +1,7 @@
 import { Client, ClientOptions } from "@graphand/client";
 import { Ora } from "ora";
 import { Config } from "./lib/Config.js";
-import { GDXDatamodels, GDXType, JSONPrimitive } from "@graphand/core";
+import { GDXDatamodels, GDXType } from "@graphand/core";
 
 declare global {
   export var client: Client | undefined;
@@ -10,13 +10,7 @@ declare global {
   export var userConfig: Config | undefined;
 }
 
-export type GDXCliType<D extends GDXDatamodels = GDXDatamodels> = {
-  "$cli.set"?: Record<string, JSONPrimitive>;
-  "$cli.file"?: Record<string, string>;
-  "$cli.function"?: Record<string, string>;
-} & Omit<GDXType<D>, "$cli.set" | "$cli.file" | "$cli.function">;
-
 export type UserConfig<D extends GDXDatamodels = {}> = {
   client: ClientOptions<D>;
-  gdx?: { path: string } | { data: GDXCliType<D> };
+  gdx?: { path: string } | { data: GDXType<D> };
 };