+++ packages/core/src/enums/job-types.ts
@@ -9,6 +9,7 @@ export enum JobTypes {
   RESTORE_SNAPSHOT = "restoreSnapshot",
   INIT_PROJECT = "initProject",
   UPGRADE_PROJECT = "upgradeProject",
+  RESTART_PROJECT = "restartProject",
   MIGRATE_PROJECT = "migrateProject",
   REMOVE_PROJECT = "removeProject",
   REMOVE_FUNCTION = "removeFunction",